requirements2.txt
scipy
numpy
matplotlib
opencv_python
tqdm
Pillow
h5py
thop
torchsummary
tensorboard